fix: migrations clean up and extensions

+++ api/database/migrations/2026_08_18_182644_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 128)->index();
             $table->text('description')->nullable();
-            $table->float('price', 2);
+            $table->decimal('price', 8, 2);
             $table->unsignedInteger('quantity');
             $table->timestamps();
         });
